WIP: input: touchscreen: add synaptics-s3908
The Synaptics S3908 is one of the TCM "on-cell" touchscreens used on the
OnePlus 8T. This driver is a port of a WIP Linux driver.

As U-Boot currently lacks any touchscreen support this is just a demo /
PoC for now. When probed it will loop forever and render the touch
coordinates on the vidconsole as well as dumping them on the serial
port.

The driver is modified to use polling as interrupts are unsupported in
U-Boot. The chip will buffer only the last touch event and associated
release event (if the finger was released) which integrates perfectly
even with a low polling rate.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>